requirements.txt
pandas
PyYAML
functions-framework==3.8.1
google-cloud-bigquery==3.25.0
requests==2.32.3
